+++ packages/apps/dev-wallet/src/pages/unlock-profile/unlock-profile.tsx
@@ -1,5 +1,4 @@
 import { AuthCard } from '@/Components/AuthCard/AuthCard';
-import { useHDWallet } from '@/modules/key-source/hd-wallet/hd-wallet';
 import {
   Avatar,
   Button,
@@ -20,7 +19,6 @@ export function UnlockProfile() {
     setError,
     formState: { isValid, errors },
   } = useForm<{ password: string }>();
-  const { unlockHDWallet } = useHDWallet();
   const { profileId } = useParams();
   const { profileList, unlockProfile } = useWallet();
   const profile = profileList.find((p) => p.uuid === profileId);
@@ -40,7 +38,6 @@ export function UnlockProfile() {
       if (!keySource) {
         throw new Error('No key source found');
       }
-      await unlockHDWallet(keySource.source, password, keySource);
     } catch (e) {
       console.log(e);
       setError('password', { type: 'manual', message: incorrectPasswordMsg });