+++ packages/apps/explorer/src/utils/graphqlIdFor.ts
@@ -1,0 +1,6 @@
+export function graphqlIdFor(type: string, hash: string): string {
+  return base64Encode(`${type}:${hash}`);
+}
+function base64Encode(str: string): string {
+  return btoa(str);
+}